refactor RESTful CRUD APIs of classroom management

+++ src/main/java/tss/repositories/ClassroomRepository.java
@@ -3,5 +3,5 @@ package tss.repositories;
 import org.springframework.data.repository.CrudRepository;
 import tss.entities.ClassroomEntity;
 
-public interface ClassroomRepository extends CrudRepository<ClassroomEntity, Short> {
+public interface ClassroomRepository extends CrudRepository<ClassroomEntity, Integer> {
 }
